Enable more build warnings for `libc-bottom-half`
While wasi-libc is compiled with `-Wall -Wextra` it ended up disabling a
lot of warnings as well to handle third-party code. This commit
reorganizes things by continuing to compile everything with `-Wall
-Wextra` but disabling lints is done exclusively for third-party code
rather than for all code in wasi-libc. In practice this means that
everything in `libc-bottom-half`, code written exclusively for this
repository, now has more lints firing. This additionally applies to
`test/src/*.c`. Much of this commit is handling these lints then and
adjusting a few patterns in places where necessary.

This commit additionally disables compiling with `-Werror` by default.
CI still enables this by default, but it means that new lints/warnings
won't be breaking for anyone else other than this repository's own CI
which should have pinned versions of Clang being used in various places.

+++ test/src/argv_two_args.c
@@ -14,7 +14,7 @@ int main(int argc, char **argv) {
 
   for (int i = 0; i < argc; i++) {
     TEST(argv[i][0] != 0, "argument should not be empty\n");
-    TEST(snprintf(buf, sizeof buf, "%s", argv[i]) < sizeof buf,
+    TEST(snprintf(buf, sizeof buf, "%s", argv[i]) < (ssize_t)sizeof buf,
          "argument is not a valid path\n");
   }
   return t_status;
